redirecionando fluxos de erro com mensagens

+++ app/Http/Controllers/MarcaController.php
@@ -55,6 +55,9 @@ class MarcaController extends Controller
     public function show($id)
     {
         $marca = $this->marca->find($id);
+        if($marca === null) {
+            return ['erro' => 'registro não encontrado'];
+        }
         return $marca;
     }
 
@@ -79,6 +82,10 @@ class MarcaController extends Controller
     public function update(Request $request, $id)
     {
         $marca = $this->marca->find($id);
+
+        if($marca === null) {
+            return ['erro' => 'registro não encontrado'];
+        }
         $marca->update($request->all());
         return $marca;
     }
@@ -92,6 +99,9 @@ class MarcaController extends Controller
     public function destroy($id)
     {
         $marca = $this->marca->find($id);
+        if($marca === null) {
+            return ['erro' => 'registro não encontrado'];
+        }
         $marca->delete();
         return ['msg' => 'A marca '."'".$marca->nome."'".' foi deletada com sucesso'];
     }
